Add files via upload
This is Vibrant University Project which is basically a University Management Project built in Semester 2 in MCA. 
This is built using Core Java, Swings, MySQL Database.
The student can register, login and sends Admission application into the college for course admission.
The Admin can see the admission application and can enroll the student. Admin is responsible for managing the students and the staff.
The student can pay their fees.